# documentación_usuarios_swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/usuarios/{id}:
    get:
      tags:
        - usuario-controller
      operationId: detalle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
        - usuario-controller
      operationId: editar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
        - usuario-controller
      operationId: eliminar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/usuarios:
    get:
      tags:
        - usuario-controller
      operationId: listar
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/Usuario'
    post:
      tags:
        - usuario-controller
      operationId: crear
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    Usuario:
      required:
        - email
        - nombre
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        fechaRegistro:
          type: string
          format: date-time