# msvc-usuarios/src/main/java/com/example/pruebaclaro/msvcusuarios/entity/Usuario.java
package com.example.pruebaclaro.msvcusuarios.entity;

import lombok.Data;

import javax.persistence.*;
import javax.validation.constraints.Email;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotEmpty;

import java.util.Date;

@Data
@Entity
@Table(name="usuarios")
public class Usuario {


    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @NotBlank
    private String nombre;

    private String apellido;

    @NotEmpty
    @Email
    @Column(unique = true)
    private String email;


    @Temporal(TemporalType.TIMESTAMP)
    private Date fechaRegistro;

    @PrePersist
    public void prePersist() {
        this.fechaRegistro = new Date();
    }

}